requirements.txt
asgiref==3.6.0
astroid==2.5.1
beautifulsoup4==4.11.2
cffi==1.15.1
colorama==0.4.4
crispy-bootstrap5==0.7
distlib==0.3.6
Django==4.1.7
django-bootstrap4==22.3
django-braces==1.15.0
django-crispy-forms==2.0
django-environ==0.10.0
django-mathfilters==1.0.0
django-shell-plus==1.1.7
filelock==3.10.2
isort==5.8.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
Pillow==9.4.0
platformdirs==3.1.1
pycparser==2.21
pylint==2.7.2
pypng==0.20220715.0
qrcode==7.4.2
soupsieve==2.4
sqlparse==0.4.3
toml==0.10.2
typing_extensions==4.5.0
tzdata==2022.7
virtualenv==20.21.0
virtualenvwrapper-win==1.2.7
whitenoise==6.4.0
wrapt==1.12.1